Bugfix + add tg_verify

+++ bot/handlers.py
@@ -1,8 +1,8 @@
+import asyncio
 import os
 import re
 import csv
 import tempfile
-import aiofiles
 
 from aiogram import types, F, Router
 from aiogram.filters import Command
@@ -29,8 +29,15 @@ def setup(router: Router, bot):
         try:
             user = await sync_to_async(CustomUser.objects.get)(email=email)
             user.telegram_id = message.from_user.id
+            user.is_tg_verified = False
             await sync_to_async(user.save)()
-            await message.answer(text.greet, reply_markup=keyboards.menu)
+            await message.answer(text.verify)
+            while True:
+                user = await sync_to_async(CustomUser.objects.get)(email=email)
+                if user.is_tg_verified:
+                    await message.answer(text.greet, reply_markup=keyboards.menu)
+                    break
+                await asyncio.sleep(5)
         except CustomUser.DoesNotExist:
             await message.answer(text.error_message_1)
 
@@ -99,44 +106,63 @@ def setup(router: Router, bot):
         await clbck.message.answer(text.ok_message, reply_markup=keyboards.exit_kb)
         
     @router.callback_query(F.data == "report")
-    async def input_parking_report(clbck: CallbackQuery):
+    async def input_report(clbck: CallbackQuery):
         user = await sync_to_async(CustomUser.objects.get)(telegram_id=clbck.from_user.id)
-        vehicle = await sync_to_async(Vehicle.objects.get)(user_id=user.id)
+        
+        vehicles = await sync_to_async(list)(Vehicle.objects.filter(user=user))
+        inline_keyboard = []
+        if vehicles:
+            for vehicle in vehicles:
+                button = [InlineKeyboardButton(text=vehicle.plate_number, callback_data=f'reports_{vehicle.id}')]
+                inline_keyboard.append(button)
+            inline_keyboard.append([InlineKeyboardButton(text="◀️ Повернутись назад", callback_data="menu")])
+            vehicle_list = InlineKeyboardMarkup(inline_keyboard=inline_keyboard)
+            await clbck.message.answer(text.report, reply_markup=vehicle_list)
+        else:
+            await clbck.message.answer(text.dont_have_car, reply_markup=keyboards.exit_kb)
+        
+    @router.callback_query(lambda car: car.data and car.data.startswith('reports_'))
+    async def process_callback_report(clbck: CallbackQuery):
+        vehicle_id = int(clbck.data.split('_')[1])
+        vehicle = await sync_to_async(Vehicle.objects.get)(id=vehicle_id)
         records = await sync_to_async(list)(ParkingSession.objects.filter(vehicle_id=vehicle.id))
+        if records:
+            all_sessions_data = []
+            for record in records:
+                payments = await sync_to_async(list)(Payment.objects.filter(parking_session_pk_id=record.id))
+                session_data = {
+                    'status': record.status,
+                    'parking_duration': record.parking_duration,
+                    'started_at': record.started_at,
+                    'end_at': record.end_at,
+                    'payments': [(payment.amount, payment.created_at) for payment in payments]
+                }
+                all_sessions_data.append(session_data)
+            
+            tmp_file_path = ''
+            with tempfile.NamedTemporaryFile(mode='w', newline='', encoding='utf-8', delete=False, suffix='.csv') as tmp_file:
+                csv_writer = csv.writer(tmp_file)
+                csv_writer.writerow(['Status', 'Parking duration', 'Started at', 'End at', 'Amount', 'Payment date'])
+                
+                for session in all_sessions_data:
+                    for payment in session['payments']:
+                        csv_writer.writerow([
+                            session['status'],
+                            session['parking_duration'],
+                            session['started_at'],
+                            session['end_at'],
+                            payment[0],
+                            payment[1]
+                        ])
+                
+                tmp_file_path = tmp_file.name
+                
+            await bot.send_document(clbck.from_user.id, FSInputFile(tmp_file_path, filename='parking_sessions.csv'))
+            os.remove(tmp_file_path)
+            await clbck.message.answer(text.csv_input, reply_markup=keyboards.exit_kb)
+        else:
+            await clbck.message.answer(text.dont_have_sessions, reply_markup=keyboards.exit_kb)
         
-        all_sessions_data = []
-        for record in records:
-            payments = await sync_to_async(list)(Payment.objects.filter(parking_session_pk_id=record.id))
-            session_data = {
-                'status': record.status,
-                'parking_duration': record.parking_duration,
-                'started_at': record.started_at,
-                'end_at': record.end_at,
-                'payments': [(payment.amount, payment.created_at) for payment in payments]
-            }
-            all_sessions_data.append(session_data)
-        
-        tmp_file_path = ''
-        with tempfile.NamedTemporaryFile(mode='w', newline='', encoding='utf-8', delete=False, suffix='.csv') as tmp_file:
-            csv_writer = csv.writer(tmp_file)
-            csv_writer.writerow(['Статус', 'Тривалість', 'Початок', 'Закінчення', 'Сума', 'Дата оплати'])
-            
-            for session in all_sessions_data:
-                for payment in session['payments']:
-                    csv_writer.writerow([
-                        session['status'],
-                        session['parking_duration'],
-                        session['started_at'],
-                        session['end_at'],
-                        payment[0],
-                        payment[1]
-                    ])
-            
-            tmp_file_path = tmp_file.name
-            
-        await bot.send_document(clbck.from_user.id, FSInputFile(tmp_file_path, filename='parking_sessions.csv'))
-        os.remove(tmp_file_path)
-        await clbck.message.answer(text.csv_input, reply_markup=keyboards.exit_kb)
